third_party/blink/web_tests/external/wpt/scroll-animations/idlharness.window-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
PASS idl_test validation
PASS Partial interface Element: member names are unique
PASS Element includes Animatable: member names are unique
PASS Element includes ParentNode: member names are unique
PASS Element includes NonDocumentTypeChildNode: member names are unique
PASS Element includes ChildNode: member names are unique
PASS Element includes Slotable: member names are unique
PASS ScrollTimeline interface: existence and properties of interface object
PASS ScrollTimeline interface object length
PASS ScrollTimeline interface object name
PASS ScrollTimeline interface: existence and properties of interface prototype object
PASS ScrollTimeline interface: existence and properties of interface prototype object's "constructor" property
PASS ScrollTimeline interface: existence and properties of interface prototype object's @@unscopables property
PASS ScrollTimeline interface: attribute scrollSource
PASS ScrollTimeline interface: attribute orientation
PASS ScrollTimeline interface: attribute startScrollOffset
PASS ScrollTimeline interface: attribute endScrollOffset
PASS ScrollTimeline interface: attribute timeRange
PASS ScrollTimeline interface: attribute fill
FAIL ScrollTimeline must be primary interface of new ScrollTimeline() assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL Stringification of new ScrollTimeline() assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL ScrollTimeline interface: new ScrollTimeline() must inherit property "scrollSource" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL ScrollTimeline interface: new ScrollTimeline() must inherit property "orientation" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL ScrollTimeline interface: new ScrollTimeline() must inherit property "startScrollOffset" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL ScrollTimeline interface: new ScrollTimeline() must inherit property "endScrollOffset" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL ScrollTimeline interface: new ScrollTimeline() must inherit property "timeRange" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
FAIL ScrollTimeline interface: new ScrollTimeline() must inherit property "fill" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "NotSupportedError: Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported"
Harness: the test ran to completion.
